Define text_object.data.l as "long long"
This ensures that text_object.data.l is at least 64 bits in size,
even in a 32-bit environment. It can cause issues if smaller,
such as text turning black instead of a desired color (see #1477).

+++ src/text_object.h
@@ -109,7 +109,7 @@ struct text_object {
     void *opaque; /* new style generic per object data */
     char *s;      /* some string */
     int i;        /* some integer */
-    long l;       /* some long integer */
+    long long l;  /* some long integer */
   } data;
 
   void *special_data;
